migrate fan wasm tests to mocha
